fix: Async read/writes.
Parser was failing because we were parsing immediately after reading the
first few bytes, and not reading the entire request until completion.
Switching to boost::beast gives us this guarantee, so now the reads will
finish before we parse the contents of the HTTP request.

Also makes quite a few changes to tests in order for them to pass, since
we now use beast requests instead of raw strings.

Change-Id: Ic5ae0ea6e1f221902387fe3b42f6e30522325aa2