skeleton files for markdown handler and testing
Markdown handler can recognize POST, GET, and DELETE, however only a plain text response indicating the request was read is given for each one.

FakeFilesystem was refactored so it could be used by multiple tests, not just CRUD. This was done so the markdown handler test didn’t have to duplicate the fake file system implementation.


Change-Id: I86f8c16e8de4a2d3356d280126c0ead55f329d8f